Moved data_manager from mapadroid/utils/ to mapadroid/
