+++ configmode.py
@@ -10,7 +10,7 @@ from mapadroid.utils.logging import initLogging, logger
 from mapadroid.utils.updater import deviceUpdater
 from mapadroid.patcher import MADPatcher
 from mapadroid.utils.walkerArgs import parseArgs
-from mapadroid.utils.data_manager import DataManager
+from mapadroid.data_manager import DataManager
 from mapadroid.websocket.WebsocketServer import WebsocketServer
 
 args = parseArgs()
