+++ mapadroid/data_manager/__init__.py
@@ -1,10 +1,11 @@
 import collections
 
 from . import modules
-from mapadroid.utils.data_manager.dm_exceptions import (
+from .dm_exceptions import (
     ModeUnknown,
     ModeNotSpecified,
-    InvalidSection
+    InvalidSection,
+    DataManagerException
 )
 
 # This is still known as the data manager but its more of a Resource Factory.  Its sole purpose is to produce a
@@ -98,8 +99,8 @@ class DataManager(object):
         if resource_def is None:
             try:
                 resource_def = self.get_resource_def(section, mode=mode)
-            except utils.data_manager.dm_exceptions.DataManagerException:
-                resource_def = copy.deepcopy(utils.data_manager.modules.MAPPINGS['area_nomode'])
+            except DataManagerException:
+                resource_def = copy.deepcopy(modules.MAPPINGS['area_nomode'])
         resources = resource_def.search(self.dbc, resource_def, self.instance_id, **params)
         results = collections.OrderedDict()
         for identifier in resources:
