+++ mapadroid/data_manager/dm_exceptions.py
