+++ mapadroid/data_manager/modules/__init__.py
