+++ mapadroid/data_manager/modules/area.py
