+++ mapadroid/data_manager/modules/area_idle.py
