+++ mapadroid/data_manager/modules/area_pokestops.py
