+++ mapadroid/data_manager/modules/auth.py
