+++ mapadroid/data_manager/modules/device.py
