+++ mapadroid/data_manager/modules/devicepool.py
