+++ mapadroid/data_manager/modules/geofence.py
