+++ mapadroid/data_manager/modules/resource.py
