+++ mapadroid/data_manager/modules/walker.py
