+++ mapadroid/data_manager/modules/walkerarea.py
