+++ mapadroid/patcher/_patch_base.py
@@ -1,5 +1,5 @@
 # Base class for a patch.  Handles the patch and basic error handling.
-from mapadroid.utils.data_manager import DataManager
+from mapadroid.data_manager import DataManager
 from mapadroid.db import DbSchemaUpdater
 from mapadroid.db.DbWrapper import DbWrapper
